Rewritten almost everything
Rewritten lots of things, the API should be more stable from now on.
The core is now a module following Doki 1.3.0 API (doki-Mowstyl 1.3.0 by default).
It can be provided when creating a QStructure and as an argument to Drewom.
Drewom is a submodule that will handle the translation from the list of operations used in a QDesign to anything another simulator may need to execute them, serving as a middleware.